requirements.txt
paho-mqtt
pyyaml
cryptography
packaging
